+++ R/qenv-constructor.R
@@ -10,13 +10,13 @@
 #'
 #' @examples
 #' new_qenv(env = list2env(list(a = 1)), code = quote(a <- 1))
-#' new_qenv(env = list2env(list(a = 1)), code = parse(text = "a <- 1"))
+#' new_qenv(env = list2env(list(a = 1)), code = parse(text = "a <- 1", keep.source = TRUE))
 #' new_qenv(env = list2env(list(a = 1)), code = "a <- 1")
 #'
 #' @return `qenv` object.
 #'
 #' @export
-setGeneric("new_qenv", function(env = new.env(parent = parent.env(.GlobalEnv)), code = expression()) standardGeneric("new_qenv")) # nolint
+setGeneric("new_qenv", function(env = new.env(parent = parent.env(.GlobalEnv)), code = character()) standardGeneric("new_qenv")) # nolint
 
 #' @rdname new_qenv
 #' @export
@@ -24,8 +24,19 @@ setMethod(
   "new_qenv",
   signature = c(env = "environment", code = "expression"),
   function(env, code) {
+    new_qenv(env, format_expression(code))
+  }
+)
+
+#' @rdname new_qenv
+#' @export
+setMethod(
+  "new_qenv",
+  signature = c(env = "environment", code = "character"),
+  function(env, code) {
     new_env <- rlang::env_clone(env, parent = parent.env(.GlobalEnv))
     lockEnvironment(new_env, bindings = TRUE)
+    if (length(code) > 0) code <- paste(code, collapse = "\n")
     id <- sample.int(.Machine$integer.max, size = length(code))
     methods::new(
       "qenv",
@@ -38,20 +49,9 @@ setMethod(
 #' @export
 setMethod(
   "new_qenv",
-  signature = c(env = "environment", code = "character"),
-  function(env, code) {
-    new_qenv(env, code = parse(text = code, keep.source = FALSE))
-  }
-)
-
-#' @rdname new_qenv
-#' @export
-setMethod(
-  "new_qenv",
   signature = c(env = "environment", code = "language"),
   function(env, code) {
-    code_expr <- as.expression(code)
-    new_qenv(env = env, code = code_expr)
+    new_qenv(env = env, code = format_expression(code))
   }
 )