+++ R/utils.R
@@ -1,41 +1,3 @@
-#' Removes leading and trailing curly brackets from character
-#' string and removes indentation of remaining contents
-#'
-#' @description `r lifecycle::badge("stable")`
-#' @param x (`character`)\cr
-#'
-#' @return character string without curly braces
-#' @keywords internal
-remove_enclosing_curly_braces <- function(x) {
-  checkmate::assert_character(x)
-  if (length(x) == 0) {
-    return(x)
-  }
-
-  open_bracket_and_spaces <- "^[[:blank:]]*\\{[[:blank:]]*$"
-  close_bracket_and_spaces <- "^[[:blank:]]*\\}[[:blank:]]*$"
-  blank_line <- "^[[:blank:]]*$"
-  four_spaces_at_start_of_line <- "^[[:blank:]]{4}"
-
-  split_text <- unlist(strsplit(x, "\n", fixed = TRUE))
-
-  # if text begins with "{   \n" and ends with "\n}   "
-  if (grepl(open_bracket_and_spaces, utils::head(split_text, 1)) &&
-    grepl(close_bracket_and_spaces, utils::tail(split_text, 1))) {
-    # remove the first and last line
-    split_text <- split_text[-c(1, length(split_text))]
-
-    # if any line is not blank then indent
-    if (!all(grepl(blank_line, split_text))) {
-      return(gsub(four_spaces_at_start_of_line, "", split_text))
-    } else {
-      return(split_text)
-    }
-  } else {
-    return(split_text)
-  }
-}
-
 #' Suppresses plot display in the IDE by opening a PDF graphics device
 #'
 #' This function opens a PDF graphics device using \code{\link[grDevices]{pdf}} to suppress
@@ -62,7 +24,30 @@ dev_suppress <- function(x) {
   force(x)
 }
 
-# converts vector of expressions to character
 format_expression <- function(code) {
-  as.character(styler::style_text(unlist(lapply(as.character(code), remove_enclosing_curly_braces))))
+  code <- lang2calls(code)
+  paste(code, collapse = "\n")
+}
+
+
+# convert language object or lists of language objects to list of simple calls
+# @param x `language` object or a list of thereof
+# @return
+# Given a `call`, an `expression`, a list of `call`s or a list of `expression`s,
+# returns a list of `calls`.
+# Symbols and atomic vectors (which may get mixed up in a list) are returned wrapped in list.
+#' @keywords internal
+lang2calls <- function(x) {
+  if (is.atomic(x) || is.symbol(x)) {
+    return(list(x))
+  }
+  if (is.call(x)) {
+    if (identical(as.list(x)[[1L]], as.symbol("{"))) {
+      as.list(x)[-1L]
+    } else {
+      list(x)
+    }
+  } else {
+    unlist(lapply(x, lang2calls), recursive = FALSE)
+  }
 }