+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.code 0.4.1.9001
+# teal.code 0.4.1.9002
 
 * The `@code` field in the `qenv` class now holds `character`, not `expression`.